// src/types/form.ts
export interface FormPayload {
  // Personal
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address: string;
  // Position
  position1: string;
  position2: string;
  position3: string;
  employmentType: string;
  workSetup: string;
  workSchedule: string;
  // Experience
  educationLevel: string;
  school: string;
  course: string;
  skillsList: string;
  tools: string;
  otherTools: string;
  // Schedule
  slot1Date: string;
  slot1Time: string;
  slot2Date: string;
  slot2Time: string;
  slot3Date: string;
  slot3Time: string;
  referralSource: string;
  referralCode: string;
  // Documents
  resumeLink: string;
  portfolioLink: string;
  videoLink: string;
}

export interface FormState {
  // Personal
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address: string;
  // Position
  position1: string;
  position2: string;
  position3: string;
  employmentType: string;
  workSetup: string[];
  workSchedule: string[];
  // Experience
  educationLevel: string;
  school: string;
  course: string;
  skills: string[];
  tools: string[];
  otherTools: string;
  // Schedule
  slot1Date: string;
  slot1Time: string;
  slot2Date: string;
  slot2Time: string;
  slot3Date: string;
  slot3Time: string;
  referralSource: string[];
  referralCode: string;
  // Documents
  resumeLink: string;
  portfolioLink: string;
  videoLink: string;
}

export const initialFormState: FormState = {
  firstName: "",
  lastName: "",
  email: "",
  phone: "",
  address: "",
  position1: "",
  position2: "",
  position3: "",
  employmentType: "",
  workSetup: [],
  workSchedule: [],
  educationLevel: "",
  school: "",
  course: "",
  skills: [],
  tools: [],
  otherTools: "",
  slot1Date: "",
  slot1Time: "",
  slot2Date: "",
  slot2Time: "",
  slot3Date: "",
  slot3Time: "",
  referralSource: [],
  referralCode: "",
  resumeLink: "",
  portfolioLink: "",
  videoLink: "",
};